requirements/base.txt
albumentations
python-multipart
fastapi==0.109.*
tortoise-orm[asyncpg]==0.20.*
aerich==0.7.2
uvicorn==0.25.0
pydantic[email]==2.3.*
pydantic-settings
httpx==0.24.1
aioboto3==12.4.0
opencv-python
onnxruntime==1.17.1
numpy==1.24.4
pandas==2.0.3
torch==2.2.2
onnx==1.16.0
scikit-image==0.23.1
scikit-learn==1.4.2
timm
loguru